requirements.txt
numpy==1.22.3
torch
torchvision
tqdm==4.64.0
opencv-python==4.6.0.66
streamlit==1.10.0
neptune-client==0.16.3
seaborn==0.11.2
matplotlib==3.5.1
kaggle
